Refactor TypeDoc configuration by removing unused helpers and consolidating options
