--- docs/src/typedocHelpers.ts
@@ -1,43 +1,0 @@
-import type { StarlightTypeDocOptions } from 'starlight-typedoc';
-
-// Utility function to create TypeDoc related paths
-export function getFilePathToPackage(name: string, path: string) {
-	return `../packages/${name}/${path}`;
-}
-
-// Utility function to create TypeDoc options for the StudioCMS packages so that each package documentation is the same when generated
-export function makeTypedocOpts(o: {
-	name: string;
-	dir: string;
-	output: string;
-	entryPoints: StarlightTypeDocOptions['entryPoints'];
-}): StarlightTypeDocOptions {
-	return {
-		tsconfig: getFilePathToPackage(o.dir, 'tsconfig.json'),
-		entryPoints: o.entryPoints,
-		output: `typedoc/${o.output}`,
-		typeDoc: {
-			plugin: [
-				'typedoc-plugin-zod',
-				'typedoc-plugin-frontmatter',
-				'./src/plugins/frontmatter.js',
-				'./src/plugins/readmes.js',
-			],
-			skipErrorChecking: true,
-			gitRemote: 'https://github.com/withstudiocms/studiocms/blob',
-			gitRevision: 'main',
-			sourceLinkTemplate:
-				'https://github.com/withstudiocms/studiocms/blob/{gitRevision}/{path}#L{line}',
-			includeVersion: true,
-			expandObjects: true,
-			expandParameters: true,
-			useCodeBlocks: true,
-			readme: 'none',
-			useHTMLAnchors: true,
-			sourceLinkExternal: true,
-			outputFileStrategy: 'modules',
-			flattenOutputFiles: true,
-			pretty: true,
-		},
-	};
-}
